+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * Debug output on Windows search for therion install info.
   * Caching Therion install path on windows on search.
   * MacOS Therion runner implementation.
+  * "Choose THConfig file and run Therion" button painted red if no Therion found.
 * Fixed bugs:
   * 'Cancel' button on settings page not closing page.
   * Therion run output should be selectable as a whole and not only single lines.